% Auto-generated: do not edit by hand
\name{navBar}

\alias{navBar}

\title{NavBar component}

\description{
Simple dark Nav Bar with mobile menu button on left.
}

\usage{
navBar(id=NULL, company=NULL, docurl=NULL, logo=NULL, navitems=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{company}{Character. The name of the client.}

\item{docurl}{Character. External google doc with documentation on using the dash for end users}

\item{logo}{Character. Sirv hosted url of the clients logo.}

\item{navitems}{Unnamed list. The Nav Bar items with their href.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

